// frontend/src/api/model/githubCom4H1RZooraInternalDomainLiveRoomMessage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainLiveRoomMessageType } from './githubCom4H1RZooraInternalDomainLiveRoomMessageType';
import type { GithubCom4H1RZooraInternalDomainUser } from './githubCom4H1RZooraInternalDomainUser';

export interface GithubCom4H1RZooraInternalDomainLiveRoomMessage {
  attachments?: number[];
  chat_id?: string;
  content?: string;
  created_at?: string;
  id?: string;
  is_edited?: boolean;
  message_type?: GithubCom4H1RZooraInternalDomainLiveRoomMessageType;
  parent_message_id?: string;
  sender?: GithubCom4H1RZooraInternalDomainUser;
  sender_id?: string;
  updated_at?: string;
}